fix: fix bugs

+++ statusserver/src/main/java/fh/bswe/statusserver/controller/StatusController.java
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import fh.bswe.statusserver.dto.StatusDto;
 import fh.bswe.statusserver.kafka.consumer.MessageConsumer;
 import fh.bswe.statusserver.kafka.producer.MessageProducer;
+import fh.bswe.statusserver.manager.SyncStatusManager;
 import fh.bswe.statusserver.service.StatusService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -13,6 +14,7 @@ import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
+@CrossOrigin(origins = "*")
 @RestController
 @RequestMapping("/status")
 public class StatusController {
@@ -22,22 +24,28 @@ public class StatusController {
     private final MessageProducer messageProducer;
     private final MessageConsumer messageConsumer;
     private final ObjectMapper objectMapper;
+    private final SyncStatusManager syncStatusManager;
 
     public StatusController(
             StatusService statusService,
             MessageProducer messageProducer,
             ObjectMapper objectMapper,
-            MessageConsumer messageConsumer) {
+            MessageConsumer messageConsumer,
+            SyncStatusManager syncStatusManager) {
         this.statusService = statusService;
         this.messageProducer = messageProducer;
         this.objectMapper = objectMapper;
         this.messageConsumer = messageConsumer;
+        this.syncStatusManager = syncStatusManager;
     }
 
     @DeleteMapping("")
     public ResponseEntity<?> deleteStatus(@RequestBody StatusDto statusDto) {
         try {
             logger.info("delete status: {}", statusDto);
+            if (syncStatusManager.isOutOfSync()) {
+                return ResponseEntity.status(HttpStatus.SERVICE_UNAVAILABLE).body("Server is out of sync.");
+            }
             statusService.deleteStatus(statusDto);
             messageProducer.sendMessage("delete", objectMapper.writeValueAsString(statusDto));
             return new ResponseEntity<>(HttpStatus.OK);
@@ -50,6 +58,9 @@ public class StatusController {
     public ResponseEntity<?> findAll() {
         try {
             logger.info("findAll");
+            if (syncStatusManager.isOutOfSync()) {
+                return ResponseEntity.status(HttpStatus.SERVICE_UNAVAILABLE).body("Server is out of sync.");
+            }
             return new ResponseEntity<>(statusService.getAllStatus(), HttpStatus.OK);
         } catch (Exception e) {
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(e.getMessage());
@@ -59,6 +70,9 @@ public class StatusController {
     @GetMapping("")
     public ResponseEntity<?> findStatusByName(@RequestParam String name) {
         try {
+            if (syncStatusManager.isOutOfSync()) {
+                return ResponseEntity.status(HttpStatus.SERVICE_UNAVAILABLE).body("Server is out of sync.");
+            }
             return new ResponseEntity<>(statusService.getStatus(name), HttpStatus.OK);
         } catch (Exception e) {
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(e.getMessage());
@@ -68,6 +82,9 @@ public class StatusController {
     @PostMapping("all")
     public ResponseEntity<?> saveAllStatus(@RequestBody List<StatusDto> status) {
         try {
+            if (syncStatusManager.isOutOfSync()) {
+                return ResponseEntity.status(HttpStatus.SERVICE_UNAVAILABLE).body("Server is out of sync.");
+            }
             statusService.setAllStatus(status);
 
             return new ResponseEntity<>(HttpStatus.CREATED);
@@ -79,6 +96,9 @@ public class StatusController {
     @PostMapping("")
     public ResponseEntity<?> saveStatus(@RequestBody StatusDto status) {
         try {
+            if (syncStatusManager.isOutOfSync()) {
+                return ResponseEntity.status(HttpStatus.SERVICE_UNAVAILABLE).body("Server is out of sync.");
+            }
             StatusDto savedStatus = statusService.setStatus(status);
             messageConsumer.setStatus(savedStatus);
             messageProducer.sendMessage("status", objectMapper.writeValueAsString(status));
@@ -88,13 +108,4 @@ public class StatusController {
         }
     }
 
-    @PostMapping("/test")
-    public ResponseEntity<?> sendMessage(@RequestParam String message) {
-        try {
-            messageProducer.sendMessage("test1", message);
-            return new ResponseEntity<>(HttpStatus.OK);
-        } catch (Exception e) {
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(e.getMessage());
-        }
-    }
 }
